Release 5.6
* This release contains breaking changes.
* Recommendations and examples for designing custom UI systems.
  * FslSimpleUI system now follows these best recommendations and comes with a default theme.
  * This includes DPI aware rendering.
* New samples
  * GLES3.RenderToTexture
  * GLES2, GLES3, Vulkan, SdfFonts
  * GLES2, GLES3, Vulkan, UI.DevNativeTexture2D
  * GLES2, GLES3, Vulkan, UI.DpiScale
  * GLES2, GLES3, Vulkan, UI.PixelPerfect
  * GLES2, GLES3, Vulkan, UI.SmoothScroll
  * GLES2, GLES3, Vulkan, UI.ThemeBasicUI
* Better CMake integration.
  * Release and debug builds now use separate output dirs by default.
  * Better support for CMake generators.
  * CMake + ninja is now the default backend for Ubuntu and Yocto for improved build speed.
  * Ubuntu builds now utilize CMake's find_package support to locate third party dependencies.
* Code modernization with PathView, StringViewLite, NumericCast and Spans.
* Clang tidy + format now requires clang 10.
* Downgraded ninja build system requirement to 1.8.2 from 1.9 for better ubuntu 18.04 compatibility.
* Better display metrics available to all samples.
* Misc: New binary texture atlas format, new bitmap font format with simple kerning.
* NativeBatch upgrades
  * SDF rendering.
  * Pixel perfect bitmap font rendering.
  * New rendering primitives.
* Now require a compiler with proper C++14 constexpr support.
  * This means GCC 5.x and below is not supported.

+++ DemoApps/Shared/FractalShader/source/Shared/FractalShader/JuliaHelper.cpp
@@ -38,6 +38,7 @@
 #include <FslBase/Math/MathHelper.hpp>
 #include <FslGraphics/Bitmap/Bitmap.hpp>
 #include <FslGraphics/Color.hpp>
+#include <FslGraphics/Render/Texture2D.hpp>
 #include <algorithm>
 #include <FslGraphics/TextureAtlas/TextureAtlasHelper.hpp>
 #include "MainAtlas.hpp"
@@ -278,8 +279,8 @@ namespace Fsl
     {
       const int val = std::min(screenResolution.X, screenResolution.Y);
       Vector2 pos = m_location;
-      pos.X *= (val * 0.5f);
-      pos.Y *= (val * 0.5f);
+      pos.X *= (static_cast<float>(val) * 0.5f);
+      pos.Y *= (static_cast<float>(val) * 0.5f);
       pos.X += (screenResolution.X * 0.5f);
       pos.Y += (screenResolution.Y * 0.5f);
       m_basic2D->Begin();
